move pull schema methods from CMManagercode to MetaPuller (#3325)
