+++ cluster/src/main/java/org/apache/iotdb/cluster/log/applier/BaseApplier.java
@@ -21,7 +21,7 @@ package org.apache.iotdb.cluster.log.applier;
 
 import org.apache.iotdb.cluster.exception.CheckConsistencyException;
 import org.apache.iotdb.cluster.log.LogApplier;
-import org.apache.iotdb.cluster.metadata.CMManager;
+import org.apache.iotdb.cluster.metadata.MetaPuller;
 import org.apache.iotdb.cluster.query.ClusterPlanExecutor;
 import org.apache.iotdb.cluster.rpc.thrift.Node;
 import org.apache.iotdb.cluster.server.member.DataGroupMember;
@@ -38,7 +38,6 @@ import org.apache.iotdb.db.qp.executor.PlanExecutor;
 import org.apache.iotdb.db.qp.physical.BatchPlan;
 import org.apache.iotdb.db.qp.physical.PhysicalPlan;
 import org.apache.iotdb.db.qp.physical.crud.InsertPlan;
-import org.apache.iotdb.db.service.IoTDB;
 import org.apache.iotdb.db.utils.SchemaUtils;
 import org.apache.iotdb.rpc.TSStatusCode;
 import org.apache.iotdb.service.rpc.thrift.TSStatus;
@@ -174,8 +173,7 @@ abstract class BaseApplier implements LogApplier {
       throws QueryProcessException {
     try {
       PartialPath path = plan.getPrefixPath();
-      ((CMManager) IoTDB.metaManager)
-          .pullTimeSeriesSchemas(Collections.singletonList(path), ignoredGroup);
+      MetaPuller.getInstance().pullTimeSeriesSchemas(Collections.singletonList(path), ignoredGroup);
     } catch (MetadataException e1) {
       throw new QueryProcessException(e1);
     }
